+++ components/project-ui/index.tsx
@@ -9,6 +9,7 @@ import ProductHeader from './ProductHeader';
 import SearchProduct from './SearchProduct';
 import FilterModal from './FilterModal';
 import ShopPagination from './ShopPagination';
+import ProductCart from './ProductCart';
 
 export {
     OrderSummary,
@@ -22,4 +23,5 @@ export {
     SearchProduct,
     FilterModal,
     ShopPagination,
+    ProductCart,
 };