+++ components/ui/index.tsx
@@ -1,4 +1,5 @@
 import Navbar from './Navbar';
 import Footer from './Footer';
+import Alert from './Alert';
 
-export { Navbar, Footer };
+export { Navbar, Footer, Alert };